refactor(tauri): consolidate legacy regen compat into client rebuild bridge
Review #4889 (4o3F): all legacy compatibility should live in the legacy
bridge. Move legacy_regen_inputs / legacy_regen_inputs_from /
regenerate_runtime_for_legacy (+ regression test) from the NyanpasuClient
facade in client/mod.rs into client/rebuild.rs next to the process-level
regen bridge, so the facade stays free of legacy-global reads.

Deliberate exception: the Config::runtime() draft write at the end of
regenerate_runtime_with stays in mod.rs — it is on the shared (new +
legacy) output path and is scoped for PR-4 (B8), not the legacy bridge.
