+++ backend/tauri/src/client/mod.rs
@@ -634,47 +634,6 @@ impl NyanpasuClient {
         self.regenerate_runtime_with(profiles, clash, app).await
     }
 
-    /// Legacy-draft snapshot -> typed build inputs for the regeneration bridge.
-    // FIXME(actor-migration): legacy-draft-aware input assembly for BC callers.
-    // Legacy Config::generate() read Config::{verge,clash}().latest() — including
-    // uncommitted drafts. Legacy side-effect writers (feat::patch_clash /
-    // patch_verge tun+service paths, CoreManager::change_core) draft first and
-    // only reseed typed actors after the mutation commits, so regenerating from
-    // typed snapshots would run one step behind (stale ports/secret/core).
-    // Convert legacy latest() via the reseed converters instead — without
-    // mutating the typed actors, so a later discard() stays a discard.
-    // New code must use rebuild_running_config()/regenerate_runtime().
-    // Remove when: PR-4/5/6 migrate the legacy writers onto typed clients.
-    fn legacy_regen_inputs() -> Result<(NyanpasuAppConfig, ClashConfig)> {
-        // MUST read latest() (draft-inclusive), never data(): legacy writers
-        // draft first and expect the regen to see it (see the FIXME above).
-        let legacy_verge = crate::config::Config::verge().latest().clone();
-        let legacy_clash = crate::config::Config::clash().latest().0.clone();
-        Self::legacy_regen_inputs_from(&legacy_verge, &legacy_clash)
-    }
-
-    /// Pure conversion half of [`Self::legacy_regen_inputs`], directly testable
-    /// without touching the process-global legacy config singletons.
-    fn legacy_regen_inputs_from(
-        legacy_verge: &crate::config::IVerge,
-        legacy_clash: &serde_yaml::Mapping,
-    ) -> Result<(NyanpasuAppConfig, ClashConfig)> {
-        let (app, _session, clash) =
-            crate::bridge::typed_config_from_legacy_parts(legacy_verge, legacy_clash)
-                .map_err(ClientError::Anyhow)?;
-        Ok((app, clash))
-    }
-
-    /// Regeneration entry for legacy bridge callers (`CoreManager::update_config`,
-    /// `feat::patch_clash`/`patch_verge` side-effect paths, `change_core`).
-    /// Profiles still come from the typed actor: their legacy writers are
-    /// rewritten against the facade in T08.
-    pub(crate) async fn regenerate_runtime_for_legacy(&self) -> Result<()> {
-        let (app, clash) = Self::legacy_regen_inputs()?;
-        let profiles = self.inner.profiles.get().await?;
-        self.regenerate_runtime_with(profiles, clash, app).await
-    }
-
     async fn regenerate_runtime_with(
         &self,
         profiles: Arc<Profiles>,
@@ -1077,35 +1036,6 @@ mod tests {
                 .expect("typed app patch should succeed");
             assert!(client.get_app_config().await.unwrap().enable_system_proxy);
         });
-    }
-
-    /// T07 review fix regression pin: the regeneration bridge assembles its
-    /// inputs from legacy verge/clash values as the writers drafted them
-    /// (feat::patch_clash / change_core draft first, reseed typed actors only
-    /// after commit). Tests the pure conversion half — the production wrapper
-    /// reads Config::{verge,clash}().latest() and must stay draft-inclusive
-    /// (mutating the process-global singletons here is inherently racy, so the
-    /// wrapper's latest() choice is locked by comment + review, not by test).
-    #[test]
-    fn legacy_regen_inputs_conversion_reflects_drafted_fields() {
-        let verge = crate::config::IVerge {
-            clash_core: Some(crate::config::nyanpasu::ClashCore::ClashRs),
-            verge_mixed_port: Some(49301),
-            ..crate::config::IVerge::default()
-        };
-        let template = crate::config::IClashTemp::template().0;
-
-        let (app, clash) = NyanpasuClient::legacy_regen_inputs_from(&verge, &template)
-            .expect("legacy regen inputs should assemble");
-        assert_eq!(
-            app.core,
-            nyanpasu_config::application::ClashCore::ClashRs,
-            "drafted clash_core must reach the app input"
-        );
-        assert_eq!(
-            clash.mixed_port.start_port, 49301,
-            "drafted mixed-port must reach the clash input"
-        );
     }
 
     #[test]
